+++ lib/animated_bottom_navigation.dart
@@ -1,4 +1,5 @@
 library animated_bottom_navigation;
 
 export 'src/item.dart';
+export 'src/tab_children_item.dart';
 export 'src/bottom_navigation.dart';